provider and model selector

+++ src/main.rs
@@ -1,10 +1,11 @@
+mod config;
+
 use crate::process::Command;
 use std::io::{self, Write};
 use async_openai::{Client, config::OpenAIConfig};
 use clap::Parser;
 use serde_json::{Value, json};
 use std::{env, fs, process};
-use std::collections::HashMap;
 
 #[derive(Parser)]
 #[command(author, version, about)]
@@ -13,55 +14,31 @@ struct Args {
     prompt: String,
 }
 
-fn get_config() -> (String, String, String) {
-    let base_urls = HashMap::from([
-        ("openrouter", vec![
-            env::var("OPENROUTER_BASE_URL").unwrap_or("error in reading OPENROUTER_BASE_URL".to_owned()),
-            env::var("OPENROUTER_API_KEY").unwrap_or("error in reading OPENROUTER_API_KEY".to_owned()),
-            String::from("nvidia/nemotron-3-nano-omni-30b-a3b-reasoning:free"),
-        ]),
-        ("free_the_ai", vec![
-            env::var("FREE_THE_AI_BASE_URL").unwrap_or("error in reading FREE_THE_AI_BASE_URL".to_owned()),
-            env::var("FREE_THE_AI").unwrap_or("error in reading FREE_THE_AI".to_owned()),
-            String::from("kai/nvidia/nemotron-3-nano-omni-30b-a3b-reasoning:free"),
-        ]),
-    ]);
-
-    loop {
-        println!("choose a base_url");
-        for i in 0..(base_urls.keys().len()) {
-            println!("{}: {}",i, base_urls.keys().nth(i).unwrap() );
-        }
-        print!("> ");
-        io::stdout().flush().unwrap();
-        let mut input = String::new(); 
-        #[allow(unused_assignments)]
-        io::stdin().read_line(&mut input).expect("failed to read line");
-
-        match input.trim().parse::<i32>() {
-            Ok(num)=>{
-                match base_urls.keys().nth(num.try_into().unwrap()) {
-                    Some(key) => {
-                        println!("{} chosen", key);
-                        return (
-                            base_urls.get(key).unwrap()[0].to_string(),
-                            base_urls.get(key).unwrap()[1].to_string(),
-                            base_urls.get(key).unwrap()[2].to_string(),
-                        );
-                    }
-                    None => {println!("number must be in the list!")}
-                }
-            }
-            Err(_e)=>{println!("invalid input!")}
-        }
-    }
-}
+static ALLOWED_CMD: [&str; 4]= [
+    "ls",
+    "pwd",
+    "cargo check",
+    "cargo check 2>&1",
+];
 
 fn create_client<T>(base_url: &str, api_key: &str) -> Client<OpenAIConfig> {
     let config = OpenAIConfig::new()
         .with_api_base(base_url)
         .with_api_key(api_key);
     Client::with_config(config)
+}
+
+fn is_file_allowed(path: String) -> bool {
+    let current_dir = std::env::current_dir().expect("Failed to get current directory");
+    let abs_path = if std::path::Path::new(&path).is_absolute() {
+        std::path::Path::new(&path).to_path_buf()
+    } else {
+        current_dir.join(&path).to_path_buf()
+    };
+    if !abs_path.starts_with(current_dir) || path.contains("..") {
+        return false;
+    }
+    true
 }
 
 async fn handle_tool_call(tool_call: &Value, messages: &mut Vec<Value>) {
@@ -72,7 +49,14 @@ async fn handle_tool_call(tool_call: &Value, messages: &mut Vec<Value>) {
     match name {
         "Read" => {
             let file_path = args["file_path"].as_str().unwrap();
-            println!("file read: {}", file_path);
+            println!(">>>> file read: {}", file_path);
+            if !is_file_allowed(file_path.to_string()){
+                eprintln!("File access not allowed '{}'", file_path);
+                messages.push(json!({
+                    "role": "tool", "tool_call_id": tool_call["id"], "content": format!("File access not allowed '{}'", file_path)
+                }));
+                return;
+            }
             match fs::read_to_string(file_path) {
                 Ok(contents) => {
                     messages.push(json!({
@@ -91,7 +75,16 @@ async fn handle_tool_call(tool_call: &Value, messages: &mut Vec<Value>) {
         "Write" => {
             let file_path = args["file_path"].as_str().unwrap();
             let cont = args["content"].as_str().unwrap();
-            println!("write tool used: {}, {}", file_path, cont);
+            println!(">>>> write tool used: {}, {}", file_path, cont);
+
+            if !is_file_allowed(file_path.to_string()){
+                eprintln!("File access not allowed '{}'", file_path);
+                messages.push(json!({
+                    "role": "tool", "tool_call_id": tool_call["id"], "content": format!("File access not allowed '{}'", file_path)
+                }));
+                return;
+            }
+
             std::fs::write(file_path, cont).unwrap();
             messages.push(json!({
                 "role": "tool", "tool_call_id": tool_call["id"], "content": cont
@@ -100,8 +93,15 @@ async fn handle_tool_call(tool_call: &Value, messages: &mut Vec<Value>) {
 
         "Bash" => {
             let cmd = args["command"].as_str().unwrap();
-            println!("shell command ran: {}", cmd);
-            let output = Command::new("powershell").arg("-c").arg(cmd).output();
+            println!(">>>> shell command ran: {}", cmd);
+            if !ALLOWED_CMD.contains(&cmd.split(" ").nth(0).unwrap()) || cmd.contains("..") {
+                eprintln!("{} is not an allowed command", cmd);
+                messages.push(json!({
+                    "role": "tool", "tool_call_id": tool_call["id"], "content": format!("{} is not an allowed command", cmd)
+                }));
+                return;
+            }
+            let output = Command::new("bash").arg("-c").arg(cmd).output();
             match &output {
                 Ok(out) => {
                     let content = String::from_utf8_lossy(&out.stdout).to_string();
@@ -163,9 +163,8 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
         }
     };
     dotenvy::dotenv().ok();
-    //let args = Args::parse();
 
-    let (base_url, api_key, model) = get_config();
+    let (base_url, api_key, model) = config::get_config();
     let client = create_client::<serde_json::Value>(&base_url, &api_key);
 
     let mut messages = vec![];
@@ -174,7 +173,6 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     while running{
         print!("> ");
         io::stdout().flush().unwrap();
-        //json!({"role": "user", "content": args.prompt});
         let mut input = String::new(); 
 
         #[allow(unused_assignments)]
@@ -184,87 +182,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
         loop {
             let response: Value = client
                 .chat()
-                .create_byot(json!({
-                    "messages": messages,
-                    //"model": "anthropic/claude-3-haiku",
-                    //"model": "kai/nvidia/nemotron-3-ultra-550b-a55b:free",
-                    "model": model,
-                    //"model": "opc/nemotron-3-ultra-free",
-                    "tools": [
-                        {
-                            "type": "function",
-                            "function": {
-                                "name": "Web",
-                                "description": "Search the web",
-                                "parameters": {
-                                    "type": "object",
-                                    "required": ["command"],
-                                    "properties": {
-                                        "command": {
-                                            "type": "string",
-                                            "description": "The phrase to search the web with"
-                                        }
-                                    }
-                                }
-                            }
-                        },
-                        {
-                            "type": "function",
-                            "function": {
-                                "name": "Bash",
-                                "description": "Execute a shell command",
-                                "parameters": {
-                                    "type": "object",
-                                    "required": ["command"],
-                                    "properties": {
-                                        "command": {
-                                            "type": "string",
-                                            "description": "The command to execute"
-                                        }
-                                    }
-                                }
-                            }
-                        },
-                        {
-                            "type": "function",
-                            "function": {
-                                "name": "Write",
-                                "description": "Write content to a file",
-                                "parameters": {
-                                    "type": "object",
-                                    "required": ["file_path", "content"],
-                                    "properties": {
-                                        "file_path": {
-                                            "type": "string",
-                                            "description": "The path of the file to write to"
-                                        },
-                                        "content": {
-                                            "type": "string",
-                                            "description": "The content to write to the file"
-                                        }
-                                    }
-                                }
-                            }
-                        },
-                        {
-                            "type": "function",
-                            "function": {
-                                "name": "Read",
-                                "description": "Read and return the contents of a file",
-                                "parameters": {
-                                    "type": "object",
-                                    "properties": {
-                                        "file_path": {
-                                            "type": "string",
-                                            "description": "The path to the file to read"
-                                        }
-                                    },
-                                    "required": ["file_path"]
-                                }
-                            }
-                        }
-                    ]
-                }))
+                .create_byot(config::create_byot(&messages, &model))
             .await?;
 
             eprintln!("Logs from your program will appear here!");
